- fixed warn: typecasting - fixed warn: empty arrays - fixed warn: no name struct - added name[]_size constant in bin2h - added some compile.cmd 'if exist' checks
* with -pedantic -fpermissive, mingw g++ seems 98% happy..
(some issues in the vstsdk 2.4)

+++ trunk/ax/parFloat.h
@@ -71,7 +71,7 @@ class parFloat : public axParameter
         if( mStep>0 )
         {
           float n = mValue * mNumSteps;
-          int st = axMin(n,(mNumSteps-1));
+          int st = (int)axMin(n,(mNumSteps-1));
           return mMin + (st*mStep);
         }
         else
